First pass at porting item architecture into the database
The way we went about this does mean we're going to manually migrate the existing inventories, but this is doable since there isn't much there yet.

There are a bunch of unused enums/classes that should be removed later, but for now we'll keep them to help with the migration.