Bumping up command timeout
I think the (very rare) issue with races getting stuck and not progressing was probably a query timeout under load. I feel comfortable letting queries run for a couple of minutes rather than the default 30 seconds. If we need to bump this up any higher it probably warrants a closer look...