Jackpots no longer use the bet multiplier
This is actually a pretty big shift in execution for slots, with most of it moving into a new class. Jackpots now return the raw jackpot amount rather than using the payout multiplier, and getting multiple complete rows of rings in one play no longer rewards multiple jackpots (e.g. millions for the first row, then 30k per row after that). A flag is set when the first complete row of rings is detected, and the jackpot amount is awarded with a message one time.

I've been meaning to do this for a while, but with the jackpot now exceeding 1 million rings, I want to avoid people winning 100 million+ when they inevitably get it due to placing high bets.